requirements.txt
aenum==3.1.16
annotated-types==0.7.0
asgiref==3.11.0
cashfree_pg==5.0.5
certifi==2026.1.4
Django==5.2
django-cors-headers==4.9.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.7
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
sentry-sdk==1.32.0
six==1.17.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.0.7
faker